pyproject.toml
[tool.poetry]
name = "contact-book"
version = "0.1.0"
description = "REST API backend for contacts book"
package-mode = false
authors = ["demagoit <33demagoit@gmail.com>"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
fastapi = "^0.110.1"
uvicorn = {extras = ["standard"], version = "^0.29.0"}
sqlalchemy = "^2.0.29"
alembic = "^1.13.1"
asyncpg = "^0.29.0"
pydentic = {extras = ["email"], version = "^0.0.1.dev3"}
python-multipart = "^0.0.9"
passlib = {extras = ["bcrypt"], version = "^1.7.4"}
python-jose = "^3.3.0"
pydantic = {extras = ["email"], version = "^2.7.0"}
python-dotenv = "^1.0.1"
pydantic-settings = "^2.2.1"
fastapi-limiter = "^0.1.6"
fastapi-mail = "^1.4.1"
cloudinary = "^1.40.0"


[tool.poetry.group.dev.dependencies]
sphinx = "^7.3.7"

[tool.poetry.group.test.dependencies]
pytest-asyncio = "^0.23.6"
pytest = "^8.2.0"
httpx = "^0.27.0"
aiosqlite = "^0.20.0"

[tool.pytest.ini_options]
testpaths = ["tests",]
pythonpath = "."

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
